requirements.txt
hbutils
hfutils
huggingface_hub
logging
numpy
opencv-python
pillow
typing


